--- pages/settings/account.vue
@@ -1,14 +1,0 @@
-<template>
-  <head>
-    <title>Impostazioni #Account • easyNet</title>
-  </head>
-
-  <SettingSelector />
-  
-</template>
-
-<script>
-  export default {
-
-  }
-</script>
