--- pages/settings/privacy.vue
@@ -1,15 +1,0 @@
-<template>
-  <head>
-    <title>Impostazioni #Privacy • easyNet</title>
-  </head>
-
-  <SettingSelector />
-  
-
-</template>
-
-<script>
-  export default {
-    
-  }
-</script>
